Added transmission message sorting
There is now a firstHopID attribute for most messages, to dictate which is the initial recipient for the transmissionThread, if it is a newly created message. Furthermore, all messages has an lastSenderID attribute, in order to determine which satellite not to send the message to.

This enables the transmission thread to sort between the types of messages and from this select the correct satellite (or groundstation) to send the message to.